requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
dataclasses-json>=0.6.0

# Phase 3 dependencies
openai>=1.0.0
numpy>=1.24.0
